Add support for the Heroku-16 stack
Like Cedar-14, Heroku-16 has stunnel pre-installed - however it's now
a new enough version that there is no need to vendor a custom build.
The pre-installed stunnel binary is already on the default `PATH`, so
supporting Heroku-16 is just a matter of skipping the vendoring step,
and manually setting the `pid` location, since the pre-installed
stunnel defaults to `/var/run/stunnel4.pid` which isn't writeable.

Support for Cedar-10 has also been dropped, since it's no longer
possible to create builds for it:
https://devcenter.heroku.com/changelog-items/755

Fixes #13.